requirements.txt
Django==3.2.24
djangorestframework==3.12.4
django-cors-headers>=3.10.0
celery>=5.2.0
redis>=4.0.0
Pillow>=8.3.0
numpy==1.23.5
psycopg2-binary>=2.9.0
gunicorn>=20.1.0
whitenoise>=5.3.0
rasterio==1.3.0
shapely==1.8.0
torch>=1.9.0
torchvision>=0.10.0
scikit-image>=0.19.0
twilio>=8.10.0
django-phonenumber-field>=7.3.0
phonenumbers>=8.13.27
opencv-python>=4.5.0

# GIS dependencies
fiona==1.9.1
geopandas==0.12.0
pyproj==3.5.0
rtree==1.0.1